fix: sperate cursor bean from job_config to batch_redis_config
